refactor(i18n): rename i18next resources file
- 【types】 rename file to i18next-resources.d.ts
- 【config】 update i18next.config.ts to reference new resourcesFile
- 【scripts】 update package.json i18n:types to format new filename
- 【code】 update import in src/@types/i18next.d.ts to match
- 【why】 standardize generated type filename and avoid dot separator
