+++ src/@types/i18next-resources.d.ts
