+++ src/@types/i18next.d.ts
@@ -1,5 +1,5 @@
 // This file is automatically generated by i18next-cli, because it was not existing. You can edit it based on your needs: https://www.i18next.com/overview/typescript#custom-type-options
-import type Resources from './i18next.resources';
+import type Resources from './i18next-resources';
 
 declare module 'i18next' {
   interface CustomTypeOptions {
